move rendering utils
