+++ packages/renderer/src/utils/RenderingUtils.ts
